Add asterisk to title when level is modified
Setting LevelHasBeenModified now has the side effect of updating the
title. I'm not sure if properties are really meant to be used this way,
but it can always be refactored later.

Also, "MLLE" in the title now comes last (like many modern programs),
and the title now uses an en dash.