Fix level corruption from invalid string encoding
This was a doozy to figure out, but straightforward to fix once I did.
I originally thought the issue had something to do with the tiles in
the level, but the help strings (specifically the ones that used '§')
were the culprits.

To summarize, the default encoding of strings in MLLE is UTF-8 but JJ2
requires Windows-1252. '§' is encoded as multiple bytes for UTF-8 which
violates assumptions in the code that all chars are encoded as one byte.

I now explicitly specify the correct encoding whenever a BinaryReader
or BinaryWriter is used, and I replaced string.ToCharArray() with
encoder.getBytes(). I also added code to automatically fix levels that
have been corrupted.