Highest event bit prevents rewording
Recreates an entirely undocumented feature of JCS's. Normally the
presence of any event on any animated tile is sufficient to make JCS
decide that word requires a brand-new entry in the dictionary; however,
that is overruled if the highest bit of that event is set. This commit
replicates that for MLLE, to avoid corrupting levels when avoidable.