Conditionally append more data to j2l, break 4 JCS
If currentPlusPropertyList says so (currently it always says no), MLLE
saves a level to have some arbitrary data after the four compressed
sections, and additionally breaks the security envelope so (unhacked)
JCS will not be able to open the level and mess it up even if it
doesn't have a password. Instead of the usual
security-envelope-breaking string, 0xBA00BE00, MLLE here uses a unique
string 0xBACABEEF that JCS reacts to in the same way but that
hypothetically allows MLLE and WebJCS to distinguish between
envelope-broken outputs (and not mess up each other's extra data) if so
desired.