Save properties to Data5 that angelscript reads
Adds a small raft of properties that are often set in or around
onLevelLoad to the PropertyGrid, then saves them inside the .j2l if at
least one of them does not equal its respective default value. These
properties can then be read and acted on from inside the .j2l by an
self-checksumming AngelScript library that is saved from MLLE, but
currently there is no code to ensure that either MLLE or JJ2+ actually
reads those properties.