Write to level script
When saving a .j2l with a Data5 section, that .j2l's .j2as will be
edited to include a call to MLLE::Setup() in the generated library
file. That included call is first removed, though, from ALL savings, in
order to avoid a script trying to include multiple versions of the
library at once.