Partial custom palette support
Partial as in it _looks_ like all the code for defining custom palettes
in Data5 is in place and should work, but there's no way to test any of
it. Generally works by passing Palette references around that may or
may not be null, and using them in place of the tileset's palette when
non-null. Unlike other properties, the palette is only written to Data5
when it is non-null, as specified by a bool byte immediately before it.