Load palettes from external files
Gets the "Load..." button in the palette editing window to work, using
all the same logic as jjPAL's load method. Unlike the jjPAL method,
though, here the palette is embedded in the .j2l instead of requiring
the entire .j2t (or other file) to be downloaded separately from J2O or
the hosting server.