Shuffle some tile counting stuff
MaxTiles and VersionNumber were only ever functions of VersionType, so
one has been made a getter and the other has been removed. TileCount
has (in name) shifted from J2TFile to J2LFile, though for the moment it
just points to J2TFile.TotalNumberOfTiles so that's not a big
difference.

(Still not directly related to data5, but trying to build up to
something here...)