Disable unused palette colors
When you look at a palette, you don't always need to look at _all_ the
colors in the palette, especially not the default windows colors. And
when remapping an image's colors, you only care about the indices used
by that particular image. As such, this commit introduces the option to
disable certain colors within a PaletteImage instance.