Call jjLayersFromLevel on self if extra tilesets
A somewhat unintuitive fix for something that JJ2+ didn't anticipate
being a problem. JJ2 only generates flipped tile images for tiles that
appear flipped in JCS, but ones that get loaded in by script (using
jjTilesFromTileset) show up too late to get included in that, which
results in flipped tiles from extra tilesets not getting drawn. Calling
jjLayersFromLevel on this level, however, doesn't add anything to the
dictionary (because there's nothing new to add) but does notify JJ2
that some tiles need to be flipped now.

Bumping the library version up to 1.1 because a few people saw 1.0
earlier while I was testing to make sure this was downloadable, and I
want to avoid any conflicting file versions, however minor.