Make layer toolstrip buttons dynamic
Instead of the 1-8 buttons in the main toolstrip being hardcoded into
MLLE, they are now regenerated every time a level is opened/created, as
well as after the layer order is (potentially) changed. (Non-default
layers get "L" instead of a number.) As far as I can tell, all buttons,
menu items, and keypresses refer to the layer I expect them to refer to.