Fix #8
This was a fun one because the problem was nowhere near where I
expected it to be. MLLE does do bounds-checking to make sure you're not
trying to place a tile outside of a layer... however, that
bounds-checking is based on the size of the layer's tile map, not the
layer itself. Normally those are the same, but not when you open a
level with a layer whose width is not a multiple of 4, because then the
tile map would get expanded a few columns for dictionary reasons. The
event map, however, was NOT so expanded, and so if you tried to place
tiles in those hidden colums, that'd be fine, unless it was layer 4, in
which case you'd be reaching outside the bounds of the event map.
Anyway, the whole thing was fixed by getting rid of the hidden columns
altogether.