Add "Edit Script..." menu option in plus props
Clicking the Edit Script option will make MLLE tell your computer to
open the level's associated .j2as file (creating it first if
necessary). When you cease to edit a level (because you close MLLE or
because you Open a different level), MLLE will check to see if a
zero-length script file was left behind by the above operation, and if
so, will delete it.