Create backup ZIP every time you save
Creates a "Backups" directory in your MLLE directory, and every time
you save a level, zips up that level (and its script and extra data
levels if any) and saves the zip in the Backups directory. When there
get to be 100 different backups, MLLE will start deleting the oldest
ones to save space.