Redo j32t file format
Instead of Data3 being expunged from the file, it's simply empty
(undefined); that should be easier to incorporate into existing code.
More to the point, a j32t (or rather a .j2t file with version 0x300)
now defines the bit depth of each tile (image) independently, so some
tiles may be 8-bit (including always tile 0) and others may be 32-bit.
A 32-bit tile is marked by setting the 0x80000000 bit in its image
address in Data1. If you are building a tileset and using two images,
one 8-bit and one 32-bit, 32-bit tile images will overrule (but not be
drawn in front of) 8-bit ones, so a tile will only be saved as 8-bit if
the equivalent area in the 32-bit image is completely empty (0/0/0/0).