Save weapons to Data5 and have AS apply them
All levels save the five custom parameter fields from the last commit
for all nine weapons. Any custom weapons write additional data to Data5
(the name, for MLLE, and any parameters) and require the use of a
slightly altered MLLE-Include library which takes an array of
MLLEWeaponApply handles (which are a shared interface).