Remove hooks as necessary
Each of the five hook functions (onMain, onPlayer, onPlayerInput,
onReceive, and onDrawAmmo) is added to the script iff at least one
weapon needs it. If a weapon previously needed a hook function, but
none of them do anymore, then the WeaponHook method call will be
removed. If this causes the hook function to have empty contents, e.g.
"void onPlayer(jjPLAYER@ player) {  }", it will be removed as well.