Smart Vflipping
just the same as the last commit but for Ctrl+I instead. Unlikely to
work very well in many scenarios, because there are a shortage of
tilesets with ceiling slopes, and most eyecandy tiles will be on the
floor (e.g. flowers), but it'd be silly not to support this at least a
little.