Smart tiles compatible with multiple tileset lvls
Sort of in first draft status right now, so there are probably some
unresolved issues, but: if a level has multiple tilesets with MLLESet
files, each of those files are loaded, and if certain smart tiles use
tiles that don't exist in the subset of tiles loaded for a specific
tileset, those tiles are ignored. If none of the default three
assignments (11,14,47) are present, a 47 has an assignment of tile 0
added just so the editor doesn't crash.