Fix multitileset Rules bug
If none of the Result tileIDs for a Rule were included in the subset of
tiles loaded for a non-primary tileset, then the Rule would be treated
as an And instead of a Then, potentially messing up everything around
it. Now such broken rules are detected and removed, along with any Ands
preceding them.