Some smart assignment logic adjustments
Modified the logic for a couple of cases that weren't producing quite
the assignments I was hoping for. Vertical tube slopes are still beaten
by horizontal tube slopes to an unfortunate extent, but if a tileset
doesn't have horizontal tubes, like the branches in Jungle, the
vertical ones work better now. Also did some more work on the logic for
when a descending slope turns into a wall, which seems to be the most
complicated of scenarios... now they have to be able to test tiles
THREE tiles above them! Even rules can't stretch that far.