Remember folders better for ZIPs
When a file is found in a folder other than the level's folder, remember its folder everywhere. Otherwise it won't be readable later if it's a script file (most likely .asc), which might lead to missing out on other files referenced by that script file.

Fixes #81. Needed to update to .NET version 4.7.2 to be able to use HashSet.TryGetValue so that was interesting, hopefully no side effects.