Offgrids saved in event map, read by JJ2
This isn't the solution I expected, but it's the one that seems to work. Or rather, the pair of solutions:

For local games, objects are created in a loop in both onLevelBegin (after the scriptwriter's had a chance to change any jjObjectPresets[*].behavior values) and onLevelReload. Once each object has finished running its STATE::START code, it gets jjOBJ::deactivates set to false (because it doesn't exist in the event map and would otherwise completely disappear once the local player left it alone) and jjOBJ::creatorID set to 1 (so it doesn't add itself into the event map when calling jjOBJ::deactivate if the local player dies in SP). creatorID is left at 0 during the actual object creation, though, in case the STATE::START code wants to check that.

For non-local games, each offgrid object has a dummy object of eventID 254 stored somewhere in the top left of the level, and then as soon as their behavior runs, they replace themselves with the actual desired offgrid objects at their desired locations. This is done in order for the server to send joining clients the status of those objects (and with consistent UIDs), usually whether they've been destroyed or not: the need to correspond to objects present in the event map when the level loads. MLLE discreetly adds these instances of eventID 254 when saving and removes them while loading, and only replaces event 0 with them.

For extra precision, it's not just eventID 254 but specifically event 0xFFFFF3FE that MLLE cares about, which is to say, eventID 254, multiplayer only difficulty, all parameters bits set. That should get rid of the events automatically in local SP, but the include library does a scan just in case it's local battle or something instead.

If a player dies in online SP, UIDs will all be thrown off-kilter, but regular events have their problems in online SP too and I don't see what can be done about it.