Use palette names for SPRITE::MAPPING layers
In MLLE's UI, use a dropdown box with the palette names in place of the generic sprite param number box. This is just to be user-friendly.

In the angelscript library, instead of the sprite param numbers saved in Data5, use whatever numbers were actually assigned to those palettes by `jjSpriteModeGetIndexMapping`. This should only make a difference in levels where `jjSpriteModeSetMapping` is called once or more before `MLLE::Setup`, e.g.

```
void onLevelLoad() {
 jjSpriteModeSetMapping(0, array<uint8>());
 MLLE::Setup();
}
```