Don't kick for offgrid weapons
JJ2+ searches the object list to see which weapons (and food) are present in the level, and kicks players for using any that aren't included in that list. MLLE's off-grid objects are created after that search. In practice, this is most commonly an issue with powerups, since often there will be only one powerup of a given weapon in a level, and if it's offgrid, players will get kicked. But hey not anymore.